+++ app/src/androidTest/java/com/epfl/drawyourpath/utils/UtilsTest.kt
@@ -2,8 +2,6 @@ package com.epfl.drawyourpath.utils
 
 
 import com.google.android.gms.maps.model.LatLng
-import com.google.mlkit.vision.digitalink.DigitalInkRecognitionModel
-import com.google.mlkit.vision.digitalink.Ink
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.`is`
 import org.junit.Assert.*
@@ -42,37 +40,5 @@ class UtilsTest {
         val ps = Utils.coordinatesToStroke(listOf(c1, c2))
 
         assertThat(ps.points, `is`(listOf(p1, p2)))
-    }
-
-    @Test
-    fun downloadModelMLDoesNotThrow() {
-        var model: DigitalInkRecognitionModel? = null
-        try {
-            model = Utils.downloadModelML().get()
-            assertNotNull(model)
-        } catch (e: Error) {
-            assertTrue(false)
-        }
-    }
-
-    @Test
-    fun modelRecognizesBox() {
-        val sideLength = 0.001
-        val baseLat = 46.5185
-        val baseLng = 6.56177
-        val c1 = LatLng(baseLat, baseLng)
-        val c2 = LatLng(baseLat + sideLength, baseLng)
-        val c3 = LatLng(baseLat + sideLength, baseLng + sideLength)
-        val c4 = LatLng(baseLat, baseLng + sideLength)
-        val c5 = LatLng(baseLat, baseLng)
-        val stroke = Utils.coordinatesToStroke(listOf(c1, c2, c3, c4, c5))
-        val ink = Ink.builder().addStroke(stroke).build()
-        val model = Utils.downloadModelML().get()
-
-        val result = Utils.recognizeDrawingML(ink, model).get()
-
-        assertNotNull(result)
-        assertEquals("box", result.classification)
-        assertEquals(1.1F, result.rawScore, 0.1F)
     }
 }
